Moved table classes out of the DbContext file and into the Domain Layer.
