+++ AnkiJapaneseFlashcardManagerTest/Anki2NoteControllerTests.cs
@@ -1,4 +1,5 @@
 ﻿using AnkiJapaneseFlashcardManager.DataAccessLayer.Contexts;
+using AnkiJapaneseFlashcardManager.DomainLayer.Entities;
 using AnkiSentenceCardBuilder.Controllers;
 using System;
 using System.Collections.Generic;
