+++ AnkiSentenceCardBuilder/AnkiJapaneseFlashcardManager.csproj
@@ -53,7 +53,6 @@
     <Folder Include="ApplicationLayer\Services\Managements\" />
     <Folder Include="DataAccessLayer\Interfaces\Contexts\" />
     <Folder Include="DataAccessLayer\Repositories\" />
-    <Folder Include="DomainLayer\Entities\" />
     <Folder Include="DomainLayer\Interfaces\Repositories\" />
   </ItemGroup>
 </Project>
