+++ AnkiSentenceCardBuilder/DataAccessLayer/Contexts/Anki2Context.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.EntityFrameworkCore;
+﻿using AnkiJapaneseFlashcardManager.DomainLayer.Entities;
+using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations.Schema;
@@ -53,68 +54,5 @@ namespace AnkiJapaneseFlashcardManager.DataAccessLayer.Contexts
                 .HasForeignKey(c => c.DeckId);
         }
 
-    }
-
-    public class Deck
-    {
-        public long Id { get; protected set; }
-        public string Name { get; protected set; }
-        [Column("mtime_secs")]
-        public int MtimeSecs { get; protected set; }
-        public byte[] Common { get; protected set; }
-        public byte[] Kind { get; protected set; }
-
-        //Navigation Properties
-        public virtual ICollection<Card> Cards { get; protected set; }
-    }
-
-    public class Card
-    {
-        public long Id { get; protected set; }
-        [Column("nid")]
-        public long NoteId { get; protected set; }
-        [Column("did")]
-        public long DeckId { get; set; }
-        [Column("ivl")]
-        public int Interval { get; protected set; }
-
-        //Navigation Properties
-        public virtual Deck Deck { get; protected set; }
-        public virtual Note Note { get; protected set; }
-
-        //Override Equals
-        public override bool Equals(object obj)
-        {
-            //Fail if not Card
-            if (obj is null || !(obj is Card otherCard))
-            {
-                return false;
-            }
-
-            return Id == otherCard.Id
-                && NoteId == otherCard.NoteId
-                && DeckId == otherCard.DeckId
-                && Interval == otherCard.Interval;
-        }
-        //Override GetHashCode
-        public override int GetHashCode()
-        {
-            return HashCode.Combine(Id, NoteId, DeckId, Interval);
-        }
-    }
-
-    public class Note
-    {
-        public long Id { get; protected set; }
-        public string Tags { get; protected set; }
-        [NotMapped] //Convert space delimited Tags field to list
-        public List<string> TagsList { get { return Tags.Split(' ').ToList(); } }
-        [Column("flds")]
-        public string Fields { get; protected set; }
-        [Column("sfld")]
-        public string SortField { get; protected set; }
-
-        //Navigation Properties
-        public virtual ICollection<Card> Cards { get; protected set; }
     }
 }
