+++ AnkiSentenceCardBuilder/Services/UpdateAnki2.cs
@@ -1,6 +1,7 @@
 using System.IO;
 using System.Threading.Tasks;
 using AnkiJapaneseFlashcardManager.DataAccessLayer.Contexts;
+using AnkiJapaneseFlashcardManager.DomainLayer.Entities;
 using AnkiSentenceCardBuilder.Controllers;
 using Azure.Storage.Blobs.Specialized;
 using Microsoft.Azure.Functions.Worker;
